# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IAM Demo API",
    "version": "2.0.0",
    "description": "API REST documentada para la demo IAM/SOC educativa."
  },
  "servers": [
    { "url": "http://localhost:3001/api" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "password": { "type": "string", "example": "Admin123!" },
          "twoFactorCode": { "type": "string" },
          "captcha": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "captcha": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Inicia sesion y crea cookie HTTPOnly",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "description": "Login correcto" },
          "401": { "description": "Credenciales, 2FA o token no validos" },
          "403": { "description": "Cuenta desactivada o bloqueada" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Registra usuario con Argon2 y politica de contrasena",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": { "201": { "description": "Usuario creado" }, "400": { "description": "Datos no validos" } }
      }
    },
...
    "/auth/profile": {
      "get": {
        "summary": "Perfil del usuario autenticado",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Perfil" }, "401": { "description": "No autenticado" } }
      }
    },
...
  }
}